third_party/WebKit/LayoutTests/fast/backgrounds/mask-box-image-width-expected.txt

This test checks that webkitMaskBoxImageWidth is correctly parsed and returned from getComputedStyle().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(testElm).webkitMaskBoxImageWidth is "auto"
PASS getComputedStyle(testElm).webkitMaskBoxImageWidth is "auto"
PASS getComputedStyle(testElm).webkitMaskBoxImageWidth is "1"
PASS getComputedStyle(testElm).webkitMaskBoxImageWidth is "0px"
PASS getComputedStyle(testElm).webkitMaskBoxImageWidth is "97%"
PASS getComputedStyle(testElm).webkitMaskBoxImageWidth is "auto"
PASS getComputedStyle(testElm).webkitMaskBoxImageWidth is "auto"
PASS successfullyParsed is true

TEST COMPLETE
